Extended grid helper functions
Added helper functions, and optional state fields, for the Grid. These will be used in the future to reduce the amount of data stored for the foreground grid (as storing a full grid of booleans for a single tetromino is wasteful). Also, it makes rotating the foreground tetromino easier as the offsets can remain the same whilst just completely replacing the tetromino data. Next thing to do is add code for collision checking using these offsets.